refactor(scripts/i18n): separate trans function from utils to deepl
