+++ .github/workflows/upload-compilation-result.yaml
@@ -28,6 +28,7 @@ jobs:
       - id: compilation
         run: |
           # Compiles the model and stores the result -- with escaped special characters.
+          yarn compile:md
           body="$(yarn compile:md 2> /dev/null | tail --lines=+3 | head --lines=-1)"
           body="${body//'%'/'%25'}"
           body="${body//$'\n'/'%0A'}"
