+++ package.json
@@ -36,6 +36,8 @@
 	"devDependencies": {
 		"cli-progress": "^3.11.2",
 		"deepl-node": "^1.7.0",
+		"dotenv": "^16.0.3",
+		"isomorphic-fetch": "^3.0.0",
 		"node-pandoc-promise": "^0.0.6",
 		"ramda": "^0.28.0",
 		"yargs": "^17.6.0"
